requirements.txt
pygame==2.5.2
numpy==1.24.3
Pillow==10.0.0
PyOpenGL==3.1.7
PyOpenGL-accelerate==3.1.7 